finally use hardhat with typescript for testing

+++ contracts/interfaces/IFizzyMigrator.sol
@@ -2,5 +2,11 @@
 pragma solidity ^0.8.9;
 
 interface IFizzyMigrator {
-    function migrate(address token, uint amountTokenMin, uint amountETHMin, address to, uint deadline) external;
+    function migrate(
+        address token,
+        uint256 amountTokenMin,
+        uint256 amountETHMin,
+        address to,
+        uint256 deadline
+    ) external;
 }
